requirements_docker.txt
msgpack
Pillow>=7.1.0
celery
pydevd-pycharm~=213.5744.248